requirements.txt
antiorm==1.2.1
click==8.1.3
db==0.1.1
db-sqlite3==0.0.1
Flask==2.1.3
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
numpy==1.24.2
pandas==1.5.3
protobuf==3.20.3
psycopg2==2.9.5
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
Werkzeug==2.2.3
